+++ remotecommands/tealium_firebase/lib/tealium_firebase.dart
@@ -1,0 +1,4 @@
+
+class TealiumFirebase {
+  static String commandName = "FirebaseRemoteCommand";
+}